+++ esperanto-tools-libs/src/KernelLaunchOptions.cpp
@@ -22,6 +22,10 @@ KernelLaunchOptions::KernelLaunchOptions() {
 KernelLaunchOptions::~KernelLaunchOptions() {
 }
 
+KernelLaunchOptions::KernelLaunchOptions(const rt::KernelLaunchOptionsImp& kOptImp) {
+  imp_ = std::make_unique<KernelLaunchOptionsImp>(kOptImp);
+}
+
 void KernelLaunchOptions::setShireMask(uint64_t shireMask) {
   setIfImpIsNull();
   imp_->shireMask_ = shireMask;
@@ -40,19 +44,21 @@ void KernelLaunchOptions::setFlushL3(bool flushL3) {
 void KernelLaunchOptions::setUserTracing(uint64_t buffer, uint32_t bufferSize, uint32_t threshold, uint64_t shireMask,
                                          uint64_t threadMask, uint32_t eventMask, uint32_t filterMask) {
   setIfImpIsNull();
-  imp_->userTraceConfig_->buffer_ = buffer;
-  imp_->userTraceConfig_->buffer_size_ = bufferSize;
-  imp_->userTraceConfig_->threshold_ = threshold;
-  imp_->userTraceConfig_->shireMask_ = shireMask;
-  imp_->userTraceConfig_->threadMask_ = threadMask;
-  imp_->userTraceConfig_->eventMask_ = eventMask;
-  imp_->userTraceConfig_->filterMask_ = filterMask;
+  UserTrace userTraceConfig{buffer, bufferSize, threshold, shireMask, threadMask, eventMask, filterMask};
+  imp_->userTraceConfig_ = userTraceConfig;
 }
 
-void KernelLaunchOptions::setStackConfig(std::byte* baseAddress, uint64_t size) {
+void KernelLaunchOptions::setStackConfig(std::byte* baseAddress, uint64_t totalSize) {
+  if ((reinterpret_cast<uint64_t>(baseAddress) % SIZE_4K) != 0) {
+    throw Exception("Stack baseAddress has to be aligned to FOUR_KB");
+  }
+  if ((totalSize % SIZE_4K) != 0) {
+    throw Exception("Stack size has to be aligned to FOUR_KB");
+  }
+
   setIfImpIsNull();
-  imp_->stackConfig_->baseAddress_ = baseAddress;
-  imp_->stackConfig_->size_ = size;
+  auto rawAddr = reinterpret_cast<uint64_t>(baseAddress);
+  imp_->stackConfig_ = StackConfiguration{rawAddr, totalSize};
 }
 
 void KernelLaunchOptions::setCoreDumpFilePath(const std::string& coreDumpFilePath) {