+++ esperanto-tools-libs/src/KernelLaunchOptionsImp.h
@@ -17,20 +17,34 @@
 
 #include "runtime/Types.h"
 
+#include <cereal/cereal.hpp>
+#include <cereal/types/optional.hpp>
+#include <cereal/types/string.hpp>
+
 namespace rt {
 
-struct stackConfiguration {
-  std::byte* baseAddress_ = nullptr;
-  uint64_t size_ = 0;
+constexpr size_t SIZE_4K = 4096;
+
+struct StackConfiguration {
+  uint64_t baseAddress_ = 0;
+  uint64_t totalSize_ = 0;
+
+  template <class Archive> void serialize(Archive& archive) {
+    archive(baseAddress_, totalSize_);
+  }
 };
 
 struct KernelLaunchOptionsImp {
-  uint64_t shireMask_ = 0xFFFFFFFF; // made dependent of ETSOC device max shire.
+  uint64_t shireMask_ = 0xFFFFFFFFUL;
   bool barrier_ = true;
   bool flushL3_ = false;
-  std::optional<UserTrace> userTraceConfig_ = std::nullopt;
-  std::string coreDumpFilePath_ = "";
-  std::optional<stackConfiguration> stackConfig_ = std::nullopt;
+  std::optional<UserTrace> userTraceConfig_;
+  std::string coreDumpFilePath_;
+  std::optional<StackConfiguration> stackConfig_;
+
+  template <class Archive> void serialize(Archive& archive) {
+    archive(shireMask_, barrier_, flushL3_, userTraceConfig_, coreDumpFilePath_, stackConfig_);
+  }
 };
 
 struct DefaultKernelOptions {