+++ esperanto-tools-libs/tests/integration-tests/test_code_loading.cpp
@@ -44,7 +44,8 @@ TEST_F(TestCodeLoading, LoadKernel) {
 }
 
 // Test loading a kernel with bss
-TEST_F(TestCodeLoading, KernelWithBss) {
+// flaky test don't execute wait [SW-19304]
+TEST_F(TestCodeLoading, DISABLED_KernelWithBss) {
   auto kernel = loadKernel("bss.elf");
   auto numElems = 150U;
   auto hSrc1 = std::vector<int>(numElems);