+++ esperanto-tools-libs/tests/unit-tests/other.cpp
@@ -22,6 +22,24 @@ TEST_F(RuntimeFixture, checkStackTraceException) {
   }
 }
 
+TEST_F(RuntimeFixture, checkStackConfigBadSizeException) {
+  // run a kernelLaunch with size not aligned to 4096B
+  constexpr size_t kTraceBytesPerHart = 4096;
+
+  KernelLaunchOptions opts;
+  std::byte* baseAddrptr = reinterpret_cast<std::byte*>((1UL << 12));
+  EXPECT_THROW(opts.setStackConfig(baseAddrptr, kTraceBytesPerHart + 5), rt::Exception);
+}
+
+TEST_F(RuntimeFixture, checkStackConfigBadBaseAddressException) {
+  // run a kernelLaunch with not aligned baseAdress pointer
+  constexpr size_t kTraceBytesPerHart = 4096;
+
+  KernelLaunchOptions opts;
+  std::byte* baseAddrptr = reinterpret_cast<std::byte*>((1UL << 12) + 15);
+  EXPECT_THROW(opts.setStackConfig(baseAddrptr, kTraceBytesPerHart), rt::Exception);
+}
+
 int main(int argc, char** argv) {
   RuntimeFixture::sDlType = RuntimeFixture::DeviceLayerImp::FAKE;
   testing::InitGoogleTest(&argc, argv);